refactor: Move GrpcMessageRouter to own file
